// public/static/charting_library/charting_library.d.ts
export declare const widget: ChartingLibraryWidgetConstructor
export declare function version(): string
export declare type AccountId = Nominal<string, "AccountID">
export declare type ActionMetaInfo = ActionDescriptionWithCallback | MenuSeparator
export declare type AvailableSaveloadVersions = "1.0" | "1.1"
// ...
export declare type HeaderWidgetButtonsMode = "fullsize" | "compact" | "adaptive"
export declare type HistoryCallback = (bars: Bar[], meta?: HistoryMetadata) => void
export declare type IBasicDataFeed = IDatafeedChartApi & IExternalDatafeed
export declare type InputFieldValidator = (value: any) => InputFieldValidatorResult
export declare type InputFieldValidatorResult =
  | PositiveBaseInputFieldValidatorResult
  | NegativeBaseInputFieldValidatorResult
export declare type LanguageCode =
  | "ar"
  | "zh"
  | "cs"
  | "da_DK"
  | "nl_NL"
  | "en"
  | "et_EE"
  | "fr"
  | "de"
  | "el"
  | "he_IL"
  | "hu_HU"
  | "id_ID"
  | "it"
  | "ja"
  | "ko"
  | "fa"
  | "pl"
  | "pt"
  | "ro"
  | "ru"
  | "sk_SK"
  | "es"
  | "sv"
  | "th"
  | "tr"
  | "vi"
  | "no"
  | "ms_MY"
  | "zh_TW"
// ...
export declare type Nominal<T, Name extends string> = T & {
  [Symbol.species]: Name
}
export declare type OnReadyCallback = (configuration: DatafeedConfiguration) => void
export declare type Order = PlacedOrder | BracketOrder
export declare type PineJS = any
export declare type QuoteData = QuoteOkData | QuoteErrorData
export declare type QuotesCallback = (data: QuoteData[]) => void
export declare type ResolutionString = Nominal<string, "ResolutionString">
// ...
export declare type StudyOverrideValueType = string | number | boolean
// ...
export declare type ThemeName = "Light" | "Dark"
export declare type TimeFrameValue = TimeFramePeriodBack | TimeFrameTimeRange
export declare type Timezone = "Etc/UTC" | CustomTimezones
// ...
export interface AccessList {
  type: "black" | "white"
  tools: AccessListItem[]
}
export interface AccessListItem {
  name: string
  grayed?: boolean
}
// ...
export interface AdditionalSymbolInfoField {
  title: string
  propertyName: string
}
// ...
export interface ChartData {
  id: string
  name: string
  symbol: string
  resolution: ResolutionString
  content: string
}
export interface ChartMetaInfo {
  id: number
  name: string
  symbol: string
  resolution: ResolutionString
  timestamp: number
}
export interface ChartingLibraryWidgetConstructor {
  new (options: ChartingLibraryWidgetOptions | TradingTerminalWidgetOptions): IChartingLibraryWidget
}
export interface ChartingLibraryWidgetOptions {
  /** @deprecated */
  container_id?: string
  container: HTMLElement | string
  datafeed: IBasicDataFeed | (IBasicDataFeed & IDatafeedQuotesApi)
  interval: ResolutionString
  symbol?: string
  auto_save_delay?: number
  autosize?: boolean
  debug?: boolean
  disabled_features?: string[]
  drawings_access?: AccessList
  enabled_features?: string[]
  fullscreen?: boolean
  height?: number
  library_path?: string
  locale: LanguageCode
  numeric_formatting?: NumericFormattingParams
  saved_data?: object
  saved_data_meta_info?: SavedStateMetaInfo
  studies_access?: AccessList
  study_count_limit?: number
  symbol_search_request_delay?: number
  timeframe?: string
  timezone?: "exchange" | Timezone
  toolbar_bg?: string
  width?: number
  charts_storage_url?: string
  charts_storage_api_version?: AvailableSaveloadVersions
  client_id?: string
  user_id?: string
  load_last_chart?: boolean
  studies_overrides?: StudyOverrides
  customFormatters?: CustomFormatters
  custom_formatters?: CustomFormatters
  overrides?: Overrides
  snapshot_url?: string
  preset?: "mobile"
  time_frames?: TimeFrameItem[]
  custom_css_url?: string
  favorites?: Favorites
  save_load_adapter?: IExternalSaveLoadAdapter
  loading_screen?: LoadingScreenOptions
  settings_adapter?: ISettingsAdapter
  theme?: ThemeName
  compare_symbols?: CompareSymbol[]
  custom_indicators_getter?: (PineJS: PineJS) => Promise<readonly CustomIndicator[]>
  additional_symbol_info_fields?: AdditionalSymbolInfoField[]
  header_widget_buttons_mode?: HeaderWidgetButtonsMode
  time_scale?: TimeScaleOptions
}
export interface ClientSnapshotOptions {
  backgroundColor: string
  font: string
  fontSize: number
  legendMode: LegendMode
  hideResolution: boolean
}
export interface CompareSymbol {
  symbol: string
  title: string
}
// ...
export interface CustomFormatter {
  format(date: Date): string
  formatLocal(date: Date): string
}
export interface CustomFormatters {
  timeFormatter: CustomFormatter
  dateFormatter: CustomFormatter
}
export interface CustomIndicator {
  readonly name: string
  readonly metainfo: any
  readonly constructor: any
}
// ...
export interface Favorites {
  intervals: ResolutionString[]
  chartTypes: string[]
}
// ...
export interface IChartingLibraryWidget {
  headerReady(): Promise<void>
  onChartReady(callback: EmptyCallback): void
  onGrayedObjectClicked(callback: (obj: GrayedObject) => void): void
  onShortcut(shortCut: string, callback: EmptyCallback): void
  subscribe<EventName extends keyof SubscribeEventsMap>(
    event: EventName,
    callback: SubscribeEventsMap[EventName]
  ): void
  unsubscribe<EventName extends keyof SubscribeEventsMap>(
    event: EventName,
    callback: SubscribeEventsMap[EventName]
  ): void
  chart(index?: number): IChartWidgetApi
  getLanguage(): LanguageCode
  setSymbol(symbol: string, interval: ResolutionString, callback: EmptyCallback): void
  remove(): void
  closePopupsAndDialogs(): void
  selectLineTool(linetool: SupportedLineTools): void
  selectedLineTool(): SupportedLineTools
  save(callback: (state: object) => void): void
  load(state: object, extendedData?: SavedStateMetaInfo): void
  getSavedCharts(callback: (chartRecords: SaveLoadChartRecord[]) => void): void
  loadChartFromServer(chartRecord: SaveLoadChartRecord): void
  saveChartToServer(
    onComplete?: EmptyCallback,
    onFail?: EmptyCallback,
    options?: SaveChartToServerOptions
  ): void
  removeChartFromServer(chartId: string, onCompleteCallback: EmptyCallback): void
  onContextMenu(callback: (unixTime: number, price: number) => ContextMenuItem[]): void
  createButton(options?: CreateButtonOptions): HTMLElement
  showNoticeDialog(params: DialogParams<() => void>): void
  showConfirmDialog(params: DialogParams<(confirmed: boolean) => void>): void
  showLoadChartDialog(): void
  showSaveAsChartDialog(): void
  symbolInterval(): SymbolIntervalResult
  mainSeriesPriceFormatter(): INumberFormatter
  getIntervals(): string[]
  getStudiesList(): string[]
  addCustomCSSFile(url: string): void
  applyOverrides<TOverrides extends StudyOverrides>(overrides: TOverrides): void
  applyStudiesOverrides(overrides: object): void
  watchList(): IWatchListApi
  activeChart(): IChartWidgetApi
  chartsCount(): number
  layout(): LayoutType
  setLayout(layout: LayoutType): void
  layoutName(): string
  changeTheme(themeName: ThemeName, options?: ChangeThemeOptions): void
  getTheme(): ThemeName
  takeScreenshot(): void
  takeClientScreenshot(options?: Partial<ClientSnapshotOptions>): Promise<HTMLCanvasElement>
  lockAllDrawingTools(): IWatchedValue<boolean>
  hideAllDrawingTools(): IWatchedValue<boolean>
  magnetEnabled(): IWatchedValue<boolean>
  magnetMode(): IWatchedValue<number>
  symbolSync(): IWatchedValue<boolean>
  intervalSync(): IWatchedValue<boolean>
  crosshairSync(): IWatchedValue<boolean>
  timeSync(): IWatchedValue<boolean>
  startFullscreen(): void
  exitFullscreen(): void
  undoRedoState(): UndoRedoState
}
export interface IDatafeedChartApi {
  getMarks?(
    symbolInfo: LibrarySymbolInfo,
    from: number,
    to: number,
    onDataCallback: GetMarksCallback<Mark>,
    resolution: ResolutionString
  ): void
  getTimescaleMarks?(
    symbolInfo: LibrarySymbolInfo,
    from: number,
    to: number,
    onDataCallback: GetMarksCallback<TimescaleMark>,
    resolution: ResolutionString
  ): void
  /**
   * This function is called if configuration flag supports_time is set to true when chart needs to know the server time.
   * The charting library expects callback to be called once.
   * The time is provided without milliseconds. Example: 1445324591. It is used to display Countdown on the price scale.
   */
  getServerTime?(callback: ServerTimeCallback): void
  searchSymbols(
    userInput: string,
    exchange: string,
    symbolType: string,
    onResult: SearchSymbolsCallback
  ): void
  resolveSymbol(
    symbolName: string,
    onResolve: ResolveCallback,
    onError: ErrorCallback,
    extension?: SymbolResolveExtension
  ): void
  getBars(
    symbolInfo: LibrarySymbolInfo,
    resolution: ResolutionString,
    periodParams: PeriodParams,
    onResult: HistoryCallback,
    onError: ErrorCallback
  ): void
  subscribeBars(
    symbolInfo: LibrarySymbolInfo,
    resolution: ResolutionString,
    onTick: SubscribeBarsCallback,
    listenerGuid: string,
    onResetCacheNeededCallback: () => void
  ): void
  unsubscribeBars(listenerGuid: string): void
  subscribeDepth?(symbol: string, callback: DomeCallback): string
  unsubscribeDepth?(subscriberUID: string): void
  getVolumeProfileResolutionForPeriod?(
    currentResolution: ResolutionString,
    from: number,
    to: number,
    symbolInfo: LibrarySymbolInfo
  ): ResolutionString
}
export interface IDatafeedQuotesApi {
  getQuotes(symbols: string[], onDataCallback: QuotesCallback, onErrorCallback: (msg: string) => void): void
  subscribeQuotes(
    symbols: string[],
    fastSymbols: string[],
    onRealtimeCallback: QuotesCallback,
    listenerGUID: string
  ): void
  unsubscribeQuotes(listenerGUID: string): void
}
// ...
export interface IExternalDatafeed {
  onReady(callback: OnReadyCallback): void
}
export interface IExternalSaveLoadAdapter {
  getAllCharts(): Promise<ChartMetaInfo[]>
  removeChart<T extends number | string>(id: T): Promise<void>
  saveChart(chartData: ChartData): Promise<number>
  getChartContent(chartId: number): Promise<string>
  getAllStudyTemplates(): Promise<StudyTemplateMetaInfo[]>
  removeStudyTemplate(studyTemplateInfo: StudyTemplateMetaInfo): Promise<void>
  saveStudyTemplate(studyTemplateData: StudyTemplateData): Promise<void>
  getStudyTemplateContent(studyTemplateInfo: StudyTemplateMetaInfo): Promise<string>
  getDrawingTemplates(toolName: string): Promise<string[]>
  loadDrawingTemplate(toolName: string, templateName: string): Promise<string>
  removeDrawingTemplate(toolName: string, templateName: string): Promise<void>
  saveDrawingTemplate(toolName: string, templateName: string, content: string): Promise<void>
}
// ...
export interface ISettingsAdapter {
  initialSettings?: InitialSettingsMap
  setValue(key: string, value: string): void
  removeValue(key: string): void
}
// ...
export interface InitialSettingsMap {
  [key: string]: string
}
// ...
export interface LoadingScreenOptions {
  foregroundColor?: string
  backgroundColor?: string
}
// ...
export interface NumericFormattingParams {
  decimal_sign: string
}
// ...
export interface Overrides {
  [key: string]: string | number | boolean
}
// ...
export interface SavedStateMetaInfo {
  uid: number
  name: string
  description: string
}
// ...
export interface StudyOverrides {
  [key: string]: StudyOverrideValueType
}
export interface StudyTemplateData {
  name: string
  content: string
}
export interface StudyTemplateMetaInfo {
  name: string
}
// ...
export interface TimeFrameItem {
  text: string
  resolution: ResolutionString
  description?: string
  title?: string
}
export interface TimeFramePeriodBack {
  type: TimeFrameType.PeriodBack
  value: string
}
export interface TimeFrameTimeRange {
  type: TimeFrameType.TimeRange
  from: number
  to: number
}
export interface TimePoint {
  time: number
}
export interface TimeScaleOptions {
  min_bar_spacing?: number
}
// ...
export interface TradingTerminalWidgetOptions extends ChartingLibraryWidgetOptions {
  brokerConfig?: SingleBrokerMetaInfo
  broker_config?: SingleBrokerMetaInfo
  restConfig?: RestBrokerConnectionInfo
  widgetbar?: WidgetBarParams
  rss_news_feed?: RssNewsFeedParams
  news_provider?: NewsProvider
  trading_customization?: TradingCustomization
  brokerFactory?(host: IBrokerConnectionAdapterHost): IBrokerWithoutRealtime | IBrokerTerminal
  broker_factory?(host: IBrokerConnectionAdapterHost): IBrokerWithoutRealtime | IBrokerTerminal
}